Restructure tests, merge embeds into messages
