+++ tests/toolbox/test_strings.py
@@ -1,6 +1,6 @@
+import datetime
 from unittest import mock
 
-import hikari
 import pytest
 
 import toolbox
@@ -39,3 +39,18 @@ async def test_fetch_message_from_link():
 
     await toolbox.fetch_message_from_link(link, bot=bot)
     bot.rest.fetch_message.assert_called_with(channel_id, message_id)
+
+
+def test_no_style():
+    time = datetime.datetime.now()
+    assert toolbox.format_dt(time) == f"<t:{int(time.timestamp())}>"
+
+
+def test_has_style():
+    time = datetime.datetime.now()
+    assert toolbox.format_dt(time, style="t") == f"<t:{int(time.timestamp())}:t>"
+
+
+def test_invalid_style():
+    with pytest.raises(ValueError):
+        toolbox.format_dt(datetime.datetime.now(), "INVALID")
